Replace butchered tiny_mce library.
The copy of tiny_mce suppied in the source distribution had to have been
butchered after build as loads of the files were only left with part of
the header comment, unclosed. This did not pass the yumicmpressor phase
in assembler.

Luckily found a compatible copy, the oldest available in the tinymce
archive but had to patch it with translations from elsewhere. Applied
some of what appears to be modification made to the sources by the
richfaces team, the proof remains in the pudding.

Replaced the minified files with source files where they were available
seems more intuitive to send those to the compressor with the added bonus
of unobfuscated sources.