Dependency Management.
Collected most of the main build modules dependencies, need to walk the
remainder of the POMs to ensure everything reflects in root and no
children are specifying version numbers.

Updated most modules to latest versions, had to roll back some.
javax-el seems to have issues with cdk and we had to remain with
the initial version.
rhino also falls down at any attempt to upgrade.