Fix a4j:support
First looked at the complex process someone attempted with registering event listeners on the parent and such jazz. Saw they went with behaviours in 4.0 and thought that’s just too bad perhaps try merge in behaviours later.

Then I thought but wait the functions and commands are working can’t we just add the event to the parent attributes and bam.

All the code was there already just not wired. Someone wanted to generate a component to ebtend tags and populate the UIComponent with the configured tag properties but that goes haywire on dependencies.

Just implemented the property popilation in the tag handler for now wdile the UIComponent populate the parent even attibute and builds the js call.