Implement InternetResource as a ResourceHandler
Richfaces 3.3.x bypasses and effectively breaks the JSF resource handling specification. We are not concerned that the components are not VDL but in the way they chose to accomplish the hack.

The current solution for discovering resources is by placing a Servlet Filter around faces and effectively intercepting requests for resources and in so doing the faces framework is unaware of these resources. Instead we should implement a ResourceHandler that can accomplish the resource resolution and allowing faces to manage InternetResources as normal faces Resources.

This patch wraps the InternetResourceBuilder as a ResourceHandler wapped to keep reference to the parent passed in by JSF. Only conflicts was the createResource contract which we refactored to createInternetResource for the existing implementation.

Created a JSF Resource wrapper of the richfaces InternetResource exposing all the contractual functionality specified for Resource. JSF also expects the Resource URL to be a qualiified java.net.URL capable of being connected to. Now of course since these are “virtual” resources and they do not reside at the physical location as descibed by URL, well not in a way URL yet knows how to obtain. So we will create a new protocol “resource” to handle the InternetResources. Created a sun.net.www.protocol.resource.Handler and exposed the InternetResource to URLConnection as needed.

Also cleaned up and sycronized the resource scanning functionality and made the ResourceBuilder more thread safe. Currently everything is still being delegated to theJSF ResourceHandler implementation and the ResourceFilter is still the goto for resource mapping but all the building blocks are in place to start switching over.